services/celery/requirements.txt
amqp==5.0.6
billiard==3.6.4.0
boto3==1.22.3
botocore==1.25.3
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
jmespath==1.0.0
kombu==5.1.0
libnfs==1.0.post4
pdf2image==1.15.1
Pillow==8.2.0
prompt-toolkit==3.0.18
pytesseract==0.3.7
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
